@@ -18,10 +18,8 @@
     "homepage": "https://github.com/cirlmcesc/laravel-swaggerdoc",
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.2.5",
-        "laravel/framework": "^7.0 || ^8.0",
-        "zircote/swagger-php": "3.*",
-        "swagger-api/swagger-ui": "^3.0",
+        "php": "^8.1",
+        "illuminate/support": "^11.0",
         "symfony/yaml": ">=3.3",
         "ext-json": "*"
     },
