@@ -1,197 +1,153 @@
 <?php
 
-return [
+use Illuminate\Support\Arr;
 
-    /*
-    |--------------------------------------------------------------------------
-    | Set Documentation API routes
-    |--------------------------------------------------------------------------
-    |
-    | Here you need to set the routing address and middleware used in the registration.
-    | When it is not needed, you can register by yourself.
-    |
-    */
+return [
 
     'auto_regist_route' => true,
 
-    'paths' => [
+    'file_directory' => 'documentations',
 
-        'data' => [
+    'backup_directory' => 'backups',
 
-            'path' => '/swaggerdoc/data',
-
-            'middlewares' => []
-
-        ],
-
-        'view' => [
-
-            'path' => '/swaggerdoc',
-
-            'middlewares' => [],
-
-        ],
-
-    ],
+    'backup_directory_nomenclature' => fn (string $previous_version): string =>
+        Arr::join(array: [now()->format('YmdHis'), $previous_version], glue: '-'),
+    
 
     /*
     |--------------------------------------------------------------------------
-    | Set file path
+    | Preset data
     |--------------------------------------------------------------------------
     |
-    | Here you need to set the path of file.
-    | Output or read will read the corresponding file.
+    | Used to generate properties set in advance for current events.
+    | OpenAPI Specification: https://swagger.io/specification/
     |
     */
 
-    'json_path' => 'documentation/api-docs.json',
+    'preset_value' => [
 
-    'yaml_path' => 'documentation/api-docs.yaml',
+        /**
+         * REQUIRED. This string MUST be the version number of the OpenAPI 
+         * Specification that the OpenAPI document uses. The openapi field SHOULD
+         * be used by tooling to interpret the OpenAPI document. This is not related
+         * to the API info.version string.
+         */
+        'openapi' => '3.1.0',
 
-
-    /*
-    |--------------------------------------------------------------------------
-    | Set openapi 3.0 json base structure
-    |--------------------------------------------------------------------------
-    |
-    | for example from coding.net.
-    |
-    |   {
-    |       "openapi": "3.0.0",
-    |       "info": {
-    |           "title": "宠物商店 API 文档",
-    |           "version": "1.0.0",
-    |           "description": "这是一篇关于宠物商店的 DEMO API 文档，仅做参考。",
-    |           "contact": {
-    |               "name": "CODING",
-    |               "url": "https://coding.net",
-    |               "email": "support@coding.net"
-    |           },
-    |           "license": {
-    |               "name": "Apache 2.0",
-    |               "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
-    |           },
-    |           "termsOfService": "http://swagger.io/terms/"
-    |       },
-    |       "servers": [
-    |           {
-    |               "url": "https://petstore.com/api/v1",
-    |               "description": "生产环境"
-    |           },
-    |           {
-    |               "url": "https://petstore.com.test.petstore.com/api/v1",
-    |               "description": "测试环境"
-    |           }
-    |       ],
-    |       "paths": [],
-    |       "components": {},
-    |       "tags": [],
-    |   }
-    |
-    */
-
-    'structure' => [
-
-        'openapi' => '3.0.0',
-
+        /**
+         * REQUIRED. Provides metadata about the API.
+         * The metadata MAY be used by tooling as required.
+         */
         'info' => [
 
-            'title' => "API 文档",
+            /** REQUIRED. The title of the API. */
+            'title' => env('APP_NAME', '') . 'API Documentation',
 
-            'version' => "1.0.0",
+            /** A short summary of the API. */
+            // "summary" => env('APP_NAME', '') . 'API Documentation Summary',
+            
+            /** 
+             * A description of the API. CommonMark syntax MAY
+             * be used for rich text representation.
+             */
+            'description' => env('APP_NAME', '') . 'API Documentation Description',
 
-            'description' => "这是关于 this project 的 API 文档",
+            /** 
+             * REQUIRED. The version of the OpenAPI document
+             * (which is distinct from the OpenAPI Specification version or the 
+             * API implementation version).
+             */
+            'version' => '0.0.1',
 
-            'contact' => [],
+            /**
+             * The contact information for the exposed API.
+             */
+            // "contact" => [
+            //     "name" => "API Support",
+            //     "url" => "http://www.example.com/support",
+            //     "email" => "support@example.com",
+            // ],
 
-            'license' => [
+            /**
+             * The license information for the exposed API.
+             */
+            // "license" => [
+            //     "name" => 'Apache 2.0',
+            //     "url" => "http://www.apache.org/licenses/LICENSE-2.0.html",
+            // ],
 
-                "name" => "Apache 2.0",
-
-                "url" => "http://www.apache.org/licenses/LICENSE-2.0.html",
-
-            ],
-
-            'termsOfService' => '',
+            /**
+             * A URL to the Terms of Service for the API. This MUST be in the form of a URL.
+             */
+            // "termsOfService" => "http://swagger.io/terms/",
 
         ],
 
+        /**
+         * An array of Server Objects, which provide connectivity information
+         * to a target server. If the servers property is not provided,
+         * or is an empty array, the default value would be a Server Object
+         * with a url value of /.
+         */
         'servers' => [
 
+            /**
+             * A single server would be described as:
+             */
             [
-
-                "url" => "https://petstore.com/api/v1",
-
-                "description" => "生产环境",
-
+                "url" => env('APP_URL', "http://development-api.example.com"),
+                "description" => env('APP_ENV', 'Development') . " server",
             ],
-
-            [
-
-                "url" => "https://petstore.com.test.petstore.com/api/v1",
-
-                "description" => "测试环境",
-
-            ],
+            // [
+            //     "url" => "https://staging-api.example.com/v1",
+            //     "description" => "Staging server"
+            // ],
+            // [
+            //     "url" => "http://api.example.com/v1",
+            //     "description" => "Production server",
+            // ],
 
         ],
 
+        /**
+         * An element to hold various schemas for the document.
+         */
+        'components' => [
+
+            'schemas' => [],
+
+            'parameters' => [],
+
+            'responses' => [],
+
+            'securitySchemes' => [],
+
+            'examples' => [],
+
+            'requestBodies' => [],
+
+            'headers' => [],
+
+            'links' => [],
+
+            'callbacks' => [],
+
+        ],
+
+        /**
+         * A list of tags used by the document with additional metadata.
+         * The order of the tags can be used to reflect on their order
+         * by the parsing tools. Not all tags that are used by theOperation Object
+         * must be declared. The tags that are not declared MAY be organized randomly
+         *  or based on the tools' logic. Each tag name in the list MUST be unique.
+         */
+        "tags" => [],
+
+        /**
+         * The available paths and operations for the API.
+         */
         'paths' => [],
 
-        'components' => [],
-
-        'tags' => [],
-
     ],
-
-
-    /*
-    |--------------------------------------------------------------------------
-    | Set standar response
-    |--------------------------------------------------------------------------
-    |
-    */
-
-    'standar_response' => [
-
-        '200' => ['description' => "Request successful"],
-
-        '401' => ['description' => "Unauthorized"],
-
-        '422' => ['description' => "The given data was invalid"],
-
-    ],
-
-
-    /*
-    |--------------------------------------------------------------------------
-    | Set default query parameters
-    |--------------------------------------------------------------------------
-    |
-    */
-
-    'query_parameters' => [
-
-        [
-            "schema" => [
-                "type" => "integer",
-                "default" => 20
-            ],
-            "in" => "query",
-            "name" => "page_size",
-            "description" => "每页返回数量"
-        ],
-
-        [
-            "schema" => [
-                "type" => "integer",
-                "default" => 1
-            ],
-            "in" => "query",
-            "name" => "page",
-            "description" => "页码"
-        ]
-
-    ]
 
 ];
